data/pyproject.toml
[project]
name = "data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.24.0",
    "azure-storage-blob>=12.26.0",
    "load-dotenv>=0.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
